clients/js: move quiet.ts --> side-effects.ts
Now that the BigInt side effect for json serialization is in this file,
it isn't just to "quiet" the log warning anymore.

Addresses review feedback from @evan-gray
