+++ clients/js/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@wormhole-foundation/wormhole-cli",
   "version": "0.0.3",
-  "sideEffects": ["src/quiet.ts"],
+  "sideEffects": ["src/side-effects.ts"],
   "description": "CLI for Wormhole related activities",
   "homepage": "https://wormhole.com",
   "bin": {
